+++ src/cs/ecl/karpAndMamidala/tawmylf/Models/WeightItem.java
@@ -26,7 +26,7 @@ public class WeightItem {
         return date;
     }
 
-    public void setDate(int utcTime) {
+    public void setDate(long utcTime) {
         this.date = new Date(utcTime);
     }
